% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaNorm-package.R
\name{TaxaNorm-datasets}
\alias{TaxaNorm-datasets}
\alias{TaxaNorm_Example_Input}
\alias{TaxaNorm_Example_Output}
\title{TaxaNorm data objects}
\usage{
data(TaxaNorm_Example_Input, package = "TaxaNorm")

data(TaxaNorm_Example_Output, package = "TaxaNorm")
}
\description{
Objects included in the TaxaNorm package, loaded with
\link[utils:data]{utils::data}
}
\section{TaxNorm_Example_Input}{


Example data
#'
}

\section{TaxNorm_Example_Output}{


Example output
}

\examples{
\dontrun{data(TaxaNorm_Example_Input, package = "TaxaNorm")
          data(TaxaNorm_Example_Output, package = "TaxaNorm")}

}
