% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R,
%   R/methods-TaxaNorm_Model_Parameters.R
\docType{class}
\name{TaxaNorm_Model_Parameters-class}
\alias{TaxaNorm_Model_Parameters-class}
\alias{TaxaNorm_Model_Parameters}
\alias{coefficients,TaxaNorm_Model_Parameters-method}
\alias{TaxaNorm_Model_Parameters-coefficients}
\alias{coefficients<-,TaxaNorm_Model_Parameters-method}
\alias{mu,TaxaNorm_Model_Parameters-method}
\alias{TaxaNorm_Model_Parameters-mu}
\alias{mu<-,TaxaNorm_Model_Parameters-method}
\alias{theta,TaxaNorm_Model_Parameters-method}
\alias{TaxaNorm_Model_Parameters-theta}
\alias{theta<-,TaxaNorm_Model_Parameters-method}
\alias{pi,TaxaNorm_Model_Parameters-method}
\alias{TaxaNorm_Model_Parameters-pi}
\alias{pi<-,TaxaNorm_Model_Parameters-method}
\title{TaxaNorm_Model_Parameters}
\usage{
TaxaNorm_Model_Parameters(coefficients, mu, theta, pi)

\S4method{coefficients}{TaxaNorm_Model_Parameters}(x)

\S4method{coefficients}{TaxaNorm_Model_Parameters}(x) <- value

\S4method{mu}{TaxaNorm_Model_Parameters}(x)

\S4method{mu}{TaxaNorm_Model_Parameters}(x) <- value

\S4method{theta}{TaxaNorm_Model_Parameters}(x)

\S4method{theta}{TaxaNorm_Model_Parameters}(x) <- value

\S4method{pi}{TaxaNorm_Model_Parameters}(x)

\S4method{pi}{TaxaNorm_Model_Parameters}(x) <- value
}
\arguments{
\item{coefficients}{Passed to \code{coefficients} slot}

\item{mu}{Passed to \code{mu} slot}

\item{theta}{Passed to \code{theta} slot}

\item{pi}{Passed to \code{pi} slot}

\item{x}{TaxaNorm_Model_Parameters object}

\item{value}{Replacement value}
}
\description{
S4 class to store TaxaNorm Parameters
}
\details{
Parameters for TaxaNorm Method
}
\section{Functions}{
\itemize{
\item \code{coefficients(TaxaNorm_Model_Parameters)}: Return \code{coefficients} slot

\item \code{mu(TaxaNorm_Model_Parameters)}: Return \code{mu} slot

\item \code{theta(TaxaNorm_Model_Parameters)}: Return \code{theta} slot

\item \code{pi(TaxaNorm_Model_Parameters)}: Return \code{pi} slot

}}
\section{Slots}{

\describe{
\item{\code{coefficients}}{\code{matrix} coefficients}

\item{\code{mu}}{\code{matrix} mu}

\item{\code{theta}}{\code{matrix} theta}

\item{\code{pi}}{\code{matrix} pi}
}}

\examples{
\dontrun{TaxaNorm_Model_Parameters(coefficients = coefficients,mu = mu,theta = theta,pi = pi)}
}
