% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxanorm_model_qc.R
\name{TaxaNorm_Model_QC}
\alias{TaxaNorm_Model_QC}
\title{Function to QC TaxNorm algorithm}
\usage{
TaxaNorm_Model_QC(TaxaNormResults)
}
\arguments{
\item{TaxaNormResults}{Input data; Results from TaxaNorm normalization}
}
\value{
a list containing qc taxnorm object
}
\description{
Function to QC TaxNorm algorithm
}
\examples{
data("TaxaNorm_Example_Output", package = "TaxaNorm")
TaxaNorm_Model_QC(TaxaNormResults = TaxaNorm_Example_Output)
}
