% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxanorm_NMDS.R
\name{TaxaNorm_NMDS}
\alias{TaxaNorm_NMDS}
\title{Function for TaxNorm NMDS}
\usage{
TaxaNorm_NMDS(TaxaNormResults, group_column)
}
\arguments{
\item{TaxaNormResults}{(Required) Input data; should be either a phyloseq object or a count matrix}

\item{group_column}{column to cluster on}
}
\value{
NMDS Plot
}
\description{
Function for TaxNorm NMDS
}
\examples{
\dontrun{TaxaNorm_NMDS(TaxaNorm_Example_Output, group_column = "body_site")}

}
