% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxanorm_input_qc.R
\name{TaxaNorm_QC_Input}
\alias{TaxaNorm_QC_Input}
\title{Function for TaxNorm input data}
\usage{
TaxaNorm_QC_Input(data)
}
\arguments{
\item{data}{(Required) Input data; should be either a phyloseq object or a count matrix}
}
\value{
QC PLots
}
\description{
Function for TaxNorm input data
}
\examples{
\dontrun{qc_data <- TaxaNorm_QC_Input(TaxaNorm_Example_Input)}

}
