% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxNorm-package.R
\name{TaxNorm-datasets}
\alias{TaxNorm-datasets}
\alias{TaxNorm_Example_Input}
\alias{TaxNorm_Example_Output}
\title{TaxNorm data objects}
\usage{
data(TaxNorm_Example_Input, package = "TaxNorm")

data(TaxNorm_Example_Output, package = "TaxNorm")
}
\description{
Objects included in the TaxNorm package, loaded with
\link[utils:data]{utils::data}
}
\section{TaxNorm_Example_Input}{


Example data
#'
}

\section{TaxNorm_Example_Output}{


Example output
}

