% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxNorm-package.R
\docType{package}
\name{TaxNorm-package}
\alias{TaxNorm}
\alias{TaxNorm-package}
\title{TaxNorm: Feature-wise Normalization for Microbiome Sequencing Data}
\description{
A novel feature-wise normalization method based on zero-inflated negative binomial model. This method assumes that the effects of sequencing depth vary for each taxon on their mean and also incorporates a rational link of zero probability and taxon dispersion as a function of sequencing depth.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wangziyue57/TaxNorm}
}

}
\author{
\strong{Maintainer}: Dillon Lloyd \email{dtlloyd@ncsu.edu}

Authors:
\itemize{
  \item Ziyue Wang \email{wangziyue57@gmail.com}
  \item Shanshan Zhao \email{shanshan.zhao@nih.gov} [contributor]
  \item Alison Motsinger-Reif \email{alison.motsinger-reif@nih.gov} [contributor]
}

}
\keyword{internal}
