% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxNorm_FitZINB.R
\name{TaxNorm_FitZINB}
\alias{TaxNorm_FitZINB}
\title{Function to run TaxNorm algorithm}
\usage{
TaxNorm_FitZINB(TaxNorm_Counts)
}
\arguments{
\item{TaxNorm_Counts}{(Required) Input TaxNorm_Counts Object}
}
\value{
a list containing the normalized count values
}
\description{
Function to run TaxNorm algorithm
}
\examples{
TaxNorm_FitZINB()

}
