% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R,
%   R/methods-TaxNorm_Model_Parameters.R
\docType{class}
\name{TaxNorm_Model_Parameters-class}
\alias{TaxNorm_Model_Parameters-class}
\alias{TaxNorm_Model_Parameters}
\title{TaxNorm_Model_Parameters}
\usage{
TaxNorm_Model_Parameters(coefficients, mu, theta, pi)
}
\arguments{
\item{coefficients}{Passed to \code{coefficients} slot}

\item{mu}{Passed to \code{mu} slot}

\item{theta}{Passed to \code{theta} slot}

\item{pi}{Passed to \code{pi} slot}
}
\description{
S4 class to store TaxNorm Parameters
}
\details{
Parameters for TaxNorm Method
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{\code{matrix} coefficients}

\item{\code{mu}}{\code{matrix} mu}

\item{\code{theta}}{\code{matrix} theta}

\item{\code{pi}}{\code{matrix} pi}
}}

