% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxnorm_NMDS.R
\name{TaxNorm_NMDS}
\alias{TaxNorm_NMDS}
\title{Function for TaxNorm NMDS}
\usage{
TaxNorm_NMDS(TaxNormResults, group_column)
}
\arguments{
\item{TaxNormResults}{(Required) Input data; should be either a phyloseq object or a count matrix}

\item{group_column}{column to cluster on}
}
\value{
NMDS Plot
}
\description{
Function for TaxNorm NMDS
}
