% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_func_ZINB.R
\name{TaxNorm_Normalization}
\alias{TaxNorm_Normalization}
\title{Function to run TaxNorm algorithm}
\usage{
TaxNorm_Normalization(
  data,
  depth = NULL,
  group = NULL,
  filter.sample.num = 10,
  filter.taxa.count = 0,
  random = TRUE,
  ncores = NULL
)
}
\arguments{
\item{data}{(Required) Input data; should be either a phyloseq object or a count matrix}

\item{depth}{sequencing depth if pre-calculated. It should be a vector with the same length and order as the column of the count data}

\item{group}{condition variables if samples are from multiple groups; should be correpsond to the column of the count data. default is NULL, where no grouping is considered}

\item{filter.sample.num, filter.taxa.count}{taxa with "filter.taxa.count" in more than "filter.sample.num" samples will be removed before testing. default is keep taxa appear in at least 10 samples within each group}

\item{random}{calculate randomized normal quantile residual}

\item{ncores}{whether multiple cores is used for parallel computing; default is max(1, detectCores() - 1)}
}
\value{
a list containing the normalized count values
}
\description{
Function to run TaxNorm algorithm
}
\examples{
TaxNorm_Normalization()

}
