% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxnorm_input_qc.R
\name{TaxNorm_QC_Input}
\alias{TaxNorm_QC_Input}
\title{Function for TaxNorm input data}
\usage{
TaxNorm_QC_Input(data)
}
\arguments{
\item{data}{(Required) Input data; should be either a phyloseq object or a count matrix}
}
\value{
QC PLots
}
\description{
Function for TaxNorm input data
}
