% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{TaxNormGenerics}
\alias{TaxNormGenerics}
\alias{rawdata}
\alias{rawdata<-}
\title{TaxNorm package generics}
\usage{
rawdata(x, ...)

rawdata(x, ...) <- value
}
\arguments{
\item{x}{TaxNorm S4 object}

\item{...}{Included for extendability; not currently used}

\item{value}{Replacement value}
}
\description{
TaxNorm package generics; see class man pages for associated
methods
}
