% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R,
%   R/methods-TaxNorm_Model_Parameters.R
\docType{class}
\name{TaxNorm_Model_Parameters-class}
\alias{TaxNorm_Model_Parameters-class}
\alias{TaxNorm_Model_Parameters}
\alias{coefficients,TaxNorm_Model_Parameters-method}
\alias{TaxNorm_Model_Parameters-coefficients}
\alias{coefficients<-,TaxNorm_Model_Parameters-method}
\alias{mu,TaxNorm_Model_Parameters-method}
\alias{TaxNorm_Model_Parameters-mu}
\alias{mu<-,TaxNorm_Model_Parameters-method}
\alias{theta,TaxNorm_Model_Parameters-method}
\alias{TaxNorm_Model_Parameters-theta}
\alias{theta<-,TaxNorm_Model_Parameters-method}
\alias{pi,TaxNorm_Model_Parameters-method}
\alias{TaxNorm_Model_Parameters-pi}
\alias{pi<-,TaxNorm_Model_Parameters-method}
\title{TaxNorm_Model_Parameters}
\usage{
TaxNorm_Model_Parameters(coefficients, mu, theta, pi)

\S4method{coefficients}{TaxNorm_Model_Parameters}(x)

\S4method{coefficients}{TaxNorm_Model_Parameters}(x) <- value

\S4method{mu}{TaxNorm_Model_Parameters}(x)

\S4method{mu}{TaxNorm_Model_Parameters}(x) <- value

\S4method{theta}{TaxNorm_Model_Parameters}(x)

\S4method{theta}{TaxNorm_Model_Parameters}(x) <- value

\S4method{pi}{TaxNorm_Model_Parameters}(x)

\S4method{pi}{TaxNorm_Model_Parameters}(x) <- value
}
\arguments{
\item{coefficients}{Passed to \code{coefficients} slot}

\item{mu}{Passed to \code{mu} slot}

\item{theta}{Passed to \code{theta} slot}

\item{pi}{Passed to \code{pi} slot}
}
\description{
S4 class to store TaxNorm Parameters
}
\details{
Parameters for TaxNorm Method
}
\section{Functions}{
\itemize{
\item \code{coefficients,TaxNorm_Model_Parameters-method}: Return \code{coefficients} slot

\item \code{mu,TaxNorm_Model_Parameters-method}: Return \code{mu} slot

\item \code{theta,TaxNorm_Model_Parameters-method}: Return \code{theta} slot

\item \code{pi,TaxNorm_Model_Parameters-method}: Return \code{pi} slot
}}

\section{Slots}{

\describe{
\item{\code{coefficients}}{\code{matrix} coefficients}

\item{\code{mu}}{\code{matrix} mu}

\item{\code{theta}}{\code{matrix} theta}

\item{\code{pi}}{\code{matrix} pi}
}}

