% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxnorm_model_qc.R
\name{TaxNorm_Model_QC}
\alias{TaxNorm_Model_QC}
\title{Function to QC TaxNorm algorithm}
\usage{
TaxNorm_Model_QC(TaxNormResults)
}
\arguments{
\item{TaxNormResults}{Input data; Results from TaxNorm normalization}
}
\value{
a list containing qc taxnorm object
}
\description{
Function to QC TaxNorm algorithm
}
