% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/methods-TaxNorm_Results.R
\docType{class}
\name{TaxNorm_Results-class}
\alias{TaxNorm_Results-class}
\alias{TaxNorm_Results}
\alias{input_data,TaxNorm_Results-method}
\alias{TaxNorm_Results-input_data}
\alias{input_data<-,TaxNorm_Results-method}
\alias{rawdata,TaxNorm_Results-method}
\alias{TaxNorm_Results-rawdata}
\alias{rawdata<-,TaxNorm_Results-method}
\alias{normdata,TaxNorm_Results-method}
\alias{TaxNorm_Results-normdata}
\alias{normdata<-,TaxNorm_Results-method}
\alias{ecdf,TaxNorm_Results-method}
\alias{TaxNorm_Results-ecdf}
\alias{ecdf<-,TaxNorm_Results-method}
\alias{model_pars,TaxNorm_Results-method}
\alias{TaxNorm_Results-model_pars}
\alias{model_pars<-,TaxNorm_Results-method}
\alias{converge,TaxNorm_Results-method}
\alias{TaxNorm_Results-converge}
\alias{converge<-,TaxNorm_Results-method}
\alias{llk,TaxNorm_Results-method}
\alias{TaxNorm_Results-llk}
\alias{llk<-,TaxNorm_Results-method}
\alias{final_df,TaxNorm_Results-method}
\alias{TaxNorm_Results-final_df}
\alias{final_df<-,TaxNorm_Results-method}
\title{TaxNorm Results}
\usage{
TaxNorm_Results(
  input_data,
  rawdata,
  normdata,
  ecdf,
  model_pars,
  converge,
  llk,
  final_df
)

\S4method{input_data}{TaxNorm_Results}(x)

\S4method{input_data}{TaxNorm_Results}(x) <- value

\S4method{rawdata}{TaxNorm_Results}(x)

\S4method{rawdata}{TaxNorm_Results}(x) <- value

\S4method{normdata}{TaxNorm_Results}(x)

\S4method{normdata}{TaxNorm_Results}(x) <- value

\S4method{ecdf}{TaxNorm_Results}(x)

\S4method{ecdf}{TaxNorm_Results}(x) <- value

\S4method{model_pars}{TaxNorm_Results}(x)

\S4method{model_pars}{TaxNorm_Results}(x) <- value

\S4method{converge}{TaxNorm_Results}(x)

\S4method{converge}{TaxNorm_Results}(x) <- value

\S4method{llk}{TaxNorm_Results}(x)

\S4method{llk}{TaxNorm_Results}(x) <- value

\S4method{final_df}{TaxNorm_Results}(x)

\S4method{final_df}{TaxNorm_Results}(x) <- value
}
\arguments{
\item{input_data}{passed to \code{input_data} slot}

\item{rawdata}{Passed to \code{rawdata} slot}

\item{normdata}{Passed to \code{normdata} slot}

\item{ecdf}{Passed to \code{ecdf} slot}

\item{model_pars}{Passed to \code{model_pars} slot}

\item{converge}{Passed to \code{converge} slot}

\item{llk}{Passed to \code{llk} slot}

\item{final_df}{Passed to \code{final_df} slot}

\item{x}{TaxNorm_Results object}

\item{value}{Replacement value}
}
\description{
S4 class to store TaxNorm Results
}
\details{
All results from the TaxNorm method and what was used to get those results
}
\section{Functions}{
\itemize{
\item \code{input_data(TaxNorm_Results)}: Return \code{input_data} slot

\item \code{rawdata(TaxNorm_Results)}: Return \code{rawdata} slot

\item \code{normdata(TaxNorm_Results)}: Return \code{normdata} slot

\item \code{ecdf(TaxNorm_Results)}: Return \code{ecdf} slot

\item \code{model_pars(TaxNorm_Results)}: Return \code{model_pars} slot

\item \code{converge(TaxNorm_Results)}: Return \code{converge} slot

\item \code{llk(TaxNorm_Results)}: Return \code{llk} slot

\item \code{final_df(TaxNorm_Results)}: Return \code{final_df} slot

}}
\section{Slots}{

\describe{
\item{\code{input_data}}{\code{ANY} phyloseq input data}

\item{\code{rawdata}}{\code{data.frame} Data frame of counts to use}

\item{\code{normdata}}{\code{data.frame} Normalized Data}

\item{\code{ecdf}}{\code{data.frame} ecdf}

\item{\code{model_pars}}{\link{TaxNorm_Model_Parameters} list of model parameters}

\item{\code{converge}}{\verb{vector(<logical>)} converge}

\item{\code{llk}}{\code{ANY} llk}

\item{\code{final_df}}{\code{ANY} final_df}
}}

\examples{
\dontrun{TaxNormResults(input_data = input_data,
                                  rawdata = rawdata,
                                   normdata = normdata,
                                   ecdf = ecdf,
                                   model_pars = model_pars,
                                   converge = converge,
                                   llk = llk,
                                   final_df = final_df)}
}
