% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/methods-TaxNorm_Results.R
\docType{class}
\name{TaxNorm_Results-class}
\alias{TaxNorm_Results-class}
\alias{TaxNorm_Results}
\alias{rawdata,TaxNorm_Results-method}
\alias{TaxNorm_Results-rawdata}
\alias{rawdata<-,TaxNorm_Results-method}
\title{TaxNorm Results}
\usage{
TaxNorm_Results(rawdata, normdata, ecdf, model_pars, converge)

\S4method{rawdata}{TaxNorm_Results}(x)

\S4method{rawdata}{TaxNorm_Results}(x) <- value
}
\arguments{
\item{rawdata}{Passed to \code{rawdata} slot}

\item{normdata}{Passed to \code{normdata} slot}

\item{ecdf}{Passed to \code{ecdf} slot}

\item{model_pars}{Passed to \code{model_pars} slot}

\item{converge}{Passed to \code{converge} slot}
}
\description{
S4 class to store TaxNorm Results
}
\details{
All results from the TaxNorm method and what was used to get those results
}
\section{Functions}{
\itemize{
\item \code{rawdata,TaxNorm_Results-method}: Return \code{rawdata} slot
}}

\section{Slots}{

\describe{
\item{\code{rawdata}}{\code{data.frame} Data frame of counts to use}

\item{\code{normdata}}{\code{data.frame} Normalized Data}

\item{\code{ecdf}}{\code{data.frame} ecdf}

\item{\code{model_pars}}{\link{TaxNorm_Model_Parameters} list of model parameters}

\item{\code{converge}}{\verb{vector(<logical>)}}
}}

