% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{TaxNormGenerics}
\alias{TaxNormGenerics}
\alias{input_data}
\alias{input_data<-}
\alias{rawdata}
\alias{rawdata<-}
\alias{normdata}
\alias{normdata<-}
\alias{ecdf}
\alias{ecdf<-}
\alias{model_pars}
\alias{model_pars<-}
\alias{converge}
\alias{converge<-}
\alias{llk}
\alias{llk<-}
\alias{final_df}
\alias{final_df<-}
\alias{coefficients}
\alias{coefficients<-}
\alias{mu}
\alias{mu<-}
\alias{theta}
\alias{theta<-}
\alias{pi}
\alias{pi<-}
\title{TaxNorm package generics}
\usage{
input_data(x, ...)

input_data(x, ...) <- value

rawdata(x, ...)

rawdata(x, ...) <- value

normdata(x, ...)

normdata(x, ...) <- value

ecdf(x, ...)

ecdf(x, ...) <- value

model_pars(x, ...)

model_pars(x, ...) <- value

converge(x, ...)

converge(x, ...) <- value

llk(x, ...)

llk(x, ...) <- value

final_df(x, ...)

final_df(x, ...) <- value

coefficients(x, ...)

coefficients(x, ...) <- value

mu(x, ...)

mu(x, ...) <- value

theta(x, ...)

theta(x, ...) <- value

pi(x, ...)

pi(x, ...) <- value
}
\arguments{
\item{x}{TaxNorm S4 object}

\item{...}{Included for extendability; not currently used}

\item{value}{Replacement value}
}
\description{
TaxNorm package generics; see class man pages for associated
methods
}
